Crate foreground_scheduler

foreground_scheduler

runs a command at interval in foreground
version: 2022.622.1243 date: 2022-06-22 author: bestia.dev repository: Github

Hashtags: #rustlang #utility #cli
My projects on Github are more like a tutorial than a finished product: bestia-dev tutorials.

screen instead of background

In Linux I love to use the screen command. In a screen session I can run a program and detach with ctrl+a, d. The program will run indefinitely. With screen -r name I can attach the session again and see what is going on. And then detach again.
Watching the stdout of the program “in foreground” is easier then reading logs. This is from the viewpoint of a developer. I want to see my program how it works after every modification.
This is great for beta web servers. They need to run indefinitely.
For other tasks like fetching data every hour I need a scheduler. The scheduler will run indefinitely inside a screen. The fetch program will run for a few seconds every hour.

Run

Run it with this arguments minute, command, args:

foreground_scheduler 4 cargo "crev repo fetch trusted"
This will run every hour at xx:04 minutes.

Development

Documentation:
https://bestia-dev.github.io/foreground_scheduler
List of prepared automation tasks for development: build, run, doc, publish,…
cargo auto

cargo crev reviews and advisory

We live in times of danger with supply chain attacks.
It is recommended to always use cargo-crev
to verify the trustworthiness of each of your dependencies.
Please, spread this info.
You can also read reviews quickly on the web:
https://web.crev.dev/rust-reviews/crates/

open-source free and free as a beer

My open-source projects are free as a beer (MIT license).
I just love programming.
But I need also to drink. If you find my projects and tutorials helpful,
please buy me a beer donating on my paypal.
You know the price of a beer in your local bar ;-)
So I can drink a free beer for your health :-)
Na zdravje! Alla salute! Prost! Nazdravlje! 🍻

Modules

scheduler_mod 🔒

scheduler_mod

Functions

enable_ansi_support

main 🔒

starting CLI